+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 0.1.3
+
+### New Features
+- Embed support
+
+### Fixes
+- Fixed display of block references inside clozes
+- Curly brackets inside code blocks and inline code are no longer parsed as cloze deletions
+
 ## 0.1.2 (2021-08-31)
 
 ### New Features